+++ examples/testing_client/src/commands/mod.rs
@@ -7,7 +7,7 @@ mod search;
 mod track;
 
 use crate::{args::Commands, save::remove_session_data};
-use tidlers::TidalClient;
+use tidlers::{TidalClient, client::models::track::config::UserUploadsInclude};
 
 pub async fn execute_command(mut tidal: TidalClient, command: Commands) -> eyre::Result<()> {
     match command {
@@ -45,7 +45,9 @@ pub async fn execute_command(mut tidal: TidalClient, command: Commands) -> eyre:
         }
 
         Commands::Uploads => {
-            let uploads = tidal.get_user_uploads(None).await?;
+            let uploads = tidal
+                .get_user_uploads(UserUploadsInclude::default(), None)
+                .await?;
             println!("{:#?}", uploads);
         }